Refactoring of the translation map.
