+++ sbomgrader/translation_maps/postprocessing/red_hat_spdx23_cdx16/spdx23.py
