+++ sbomgrader/translation_maps/red_hat_spdx23_cdx16.yml
