+++ sbomgrader/translation_maps/transformers/red_hat_spdx23_cdx16/utils.py
