+++ tests/test_translation.py
@@ -1,4 +1,3 @@
-import json
 from typing import Any
 
 import pytest
@@ -13,7 +12,7 @@ from sbomgrader.core.utils import get_mapping
 @pytest.fixture(scope="session")
 def built_in_translation_map() -> TranslationMap:
     yield TranslationMap.from_file(
-        "sbomgrader/translation_maps/sample_spdx23_cdx16.yml"
+        "sbomgrader/translation_maps/red_hat_spdx23_cdx16.yml"
     )
