+++ tests/testdata/test_translation/actual_data/openssl-3.0.7-18.el9_2.spdx.json
